Fix some issues with indenting & unindenting
The `indentation_level` was not saved/restored correctly when adding all
items in one go. The problem did not occur with items loaded from disk,
because `load_from_string` did contain code that was skipped otherwise.
Now that code was moved to the `add` and `remove` methods instead.

Also, adds some new `emit` flags to both methods that defaults to `true`
and determines whether the `changed` will be emitted at the end or not.
This allows sending a more precise reason for the change in `indent` and
`unindent`. Previously, when indenting a to-do that first printed "to-do
removed", then "to-do added". Now it only prints "to-do intended".

Further removes a condition from the `indente_todo` and `unindent_todo`
methods that required the item to already have a text before submitting
the change to the `Data` singleton, which resulted in disparities when
indenting the item first and only adding the text afterwards.

Lastly, makes sure to remember the caret position before reparenting the
to-dos, so it can be restored when editing the (un)indented to-do again.
For that, a `remember_caret_position` method was added.