Include extension in `Settings.data_format_save`
Previously, in all places but one (which is also destined to be removed
later) the ".txt"-extension was added manually anyways.

Also, adds a `has_file` method to the `Data` singleton that's supposed
to be used as a drop-in replacement for calling `FileAccess.file_exists`
in the calendar widget. It will return true if there's any (non-empty)
FileData with the given filename, even if it's not yet written to disk.