Adapt global search feature to use `Data`
Note that this only concern the global search itself, the "Saved Search"
part is still using `Cache`. Migrating was fairly straightforward, but
required a few changes elsewhere to make it work properly:

- The `save_to_disk` method in `FileData` previously didn't consider the
  fact that a file's to-do list might be empty and should not be saved.
- The `reload` method, called after `NOTIFICATION_APPLICATION_FOCUS_IN`,
  did effectively undo all modifications of the `files` dictionary via
  the `get_file` method, as it was unloading every file without a file.
  As a bandaid fix, if a file is created that way, its `is_preliminary`
  variable will be set to `true`, meaning it won't be unloaded. Saving
  the file to disk will reset `is_preliminary` back to `false`.

The new approach no longer requires a deferred connection (to give the
Cache time to update) or parsing the raw string anew each time.

Also, this commit removes all explicit mentions of `Data` in `data.gd`
itself, for the sake of brevity and conciseness.