Fix several issues with yesterday's commit
Turns out, the changes made in 0ad874d were problematic for various
reasons. Most importantly, passing the same ToDoData to another list
would add the *same* data object in multiple places at once! That's why
this commit introduces a custom `duplicate()` method to get an actual,
independenten copy of the ToDoData it was called for.

Also, connecting the new `to_do_added` signal to the `add_todo` method
meant that on several occasions *two* new to-dos did get created. To
prevent this, all places where this was the case now directly call the
`add` method on the data instead of calling `add_todo`.
To make drag & drop operations update the list view properly (and not
just the underlying data), this commit also introduces a `to_do_removed`
signal and changes the respective `_drop_data` method in such a way that
the ToDoItem node is no longer moved, but removed and added.

Lastly, this commit fixes the `last_modified` variable not being updated
when calling FileData's `save_to_disk` method, which resulted in strange
results when e.g. adding a new to-do, and then moving the mouse cursor
out off the app and back into again which (falsely!) triggered a reload.