Add `Data.changed` signal
... which can be used to retrigger the search for the last mention of a
saved search query. Note that this naive solution will trigger on *any*
changes and doesn't try to narrow it down to changes that might actually
*affect* a given saved search! Thus, it can certainly be improved upon,
but (at least on my machine) the search is fast enough to postpone that
optimization to another day and embrace the simpler implementation. The
new solution also doesn't require the `cache_key` variable anymore!