Adapt review view to use `Data`
This required adding a new `review_date` variable to `ToDoData`, which
will equal `null` if it ain't set. I also added a special condition to
the `_update_mouse_cursor_state` method in `BetterButton` to prevent an
error when trying to access the viewport after the button was removed
from the scene tree (which is the case for the calendar widget).

Also, fixes the calendar widget not being hidden by default.