Fix global search ignoring all sub items
The previous code was only looping through the topmost to-do list of a
file, ignoring the fact that any of those to-dos might have sub to-dos.

This commit adds a `search` method to `ToDoListData`, that recursively
searches through all of its to-dos (and their sub to-dos, and their sub
to-dos sub to-dos and so on) and returns an array of `ToDoData` matching
the given query. `ToDoData` has an `edit` method now, which will first
change `Settings.current_date` and then emit an `edit_requested` signal
to prompt the corresponding `ToDoItem` to switch into edit mode.