Fix some bugs when searching to-dos
First, when typing text into the search bar while the capture list panel
wasn't visible, the list's scroll button would throw a null error. Thus,
the `_update_button` method of the ScrollButton now returns early if the
button isn't currently visible in the tree.

Second, the global search results wouldn't update after modifying a file
outside of Habituary. This is now covered by re-running `search()` after
the `Data.changed` signal is emitted. Originally, the code also checked
if the key of the changed file matched the date of a search result group
before running the method, but that didn't consider the possibility of a
file previously not matching a query changing to matching it now!

Third, to make all of this work, the `Data.changed` signal is now also
emitted after reloading a file that was modified outside the app.