Reload `todo_list` if modified outside the app
Previously, only the underlying `FileData` was updated, not its visual
representation in the list view panel! This commit adds a `reloaded`
signal, which prompts the associated DayPanel to free all of the old
items and create new ones based on the given data.
There might be a (much more complex) solution that keeps to-dos that
haven't changed intact, but since I don't expect many people to make
changes to a file outside the app while the app is running, it didn't
really seem worth all that extra effort.

Also, removes two lines of code I had commented out, but then forgot to
remove in the last commit.