Delete empty files when scrolling the list view
Each loaded DayPanel will create a preliminary `FileData` object for its
date if there isn't one. Previously these objects did *not* get removed
once the corresponding date was no longer present in the list view!

This commit changes that by adding a `delete_from_disk` method that will
both delete the corresponding file on disk (if there is any) and unload
the object from the `Data.files` dictionary (if it is in there).