Prevent FileData in list view from being unloaded
... by setting the new `keep_loaded` variable to `true` when loading the
data in `todo_list.gd` and only setting it back to `false` once the list
has exited the tree again (i.e. is no longer visible).
This replaces the old `is_preliminary` variable, which tackled a similar
issue but didn't work for all possible cases: When deleting *the last*
to-do in a list and then moving the mouse out of the window before that
change was saved to disk and then back into it again, the data of that
file would get unloaded and was no longer considered by global searches.