Clean up `DayPanel` class
Most of these changes are cosmetic, like renaming things or changing
their order. Also, the new code makes use of the `Date.equals` method
as well as the new `left_mouse_button` input action now.

The most notable change is that all methods that use the view mode in
some way will now consider the view mode cap as well, so they will keep
working correctly even when the window is resized down.

Also, methods will no longer check if `date` is set or not. The variable
being immutable plus the assertion in ready should guarantee that!

Lastly, I no longer consider the way the Header's drag forwarding is
handled "pretty hacky", so I removed the "FIXME" for that. With the
inclusion of `Utils.MAX_INT` this seems like a decent solution!