Add `caret_position` variable to `ToDoData`
Its value does deliberately NOT (!) persist across restarts. It gets
automatically set when a to-do's text is submitted. Alternatively, it
can be set by calling the `edit` method with an optional argument.

Also, calling `edit` will no longer center the view on the date of the
file with the to-do, if its date is already visible in the list view.