Fix some issues with the new `caret_position`
Firstly, `item_list.gd` still contained two `remember_caret_position`
calls which resulted in an error when (un-)indenting a to-do.

Secondly, the `at_caret_position` argument of `ToDoData.edit` defaults
to `null` instead of 0 now, no longer resetting it on an edit.

Lastly, the `caret_position` is now updated in `_on_edit_focus_exited`,
i.e. even if it regains the focus one frame later (which is the case
when indenting or unindenting a to-do).